move puzzle and memory game to their sepperate folders
