+++ src/games/games.html
@@ -1,42 +1,47 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Lojërat</title>
-    <link rel="stylesheet" href="games.css">
-
-</head>
-<body>
+  </head>
+  <body>
     <section class="hero">
-        <h1><i class="fa-solid fa-gamepad"></i> Luaj dhe mëso me Duartë</h1>
-        <p>
-            Zbavituni duke mësuar gjuhën e shenjave përmes lojërave argëtuese! Provoni lojërat tona për të përmirësuar kujtesën dhe për të zbuluar më shumë rreth gjesteve në mënyrë kreative.
-        </p>
-      </section>
+      <h1><i class="fa-solid fa-gamepad"></i> Luaj dhe mëso me Duartë</h1>
+      <p>
+        Zbavituni duke mësuar gjuhën e shenjave përmes lojërave argëtuese!
+        Provoni lojërat tona për të përmirësuar kujtesën dhe për të zbuluar më
+        shumë rreth gjesteve në mënyrë kreative.
+      </p>
+    </section>
     <div class="all-cont">
-      
-
-        <div class="game-container">
-            <!-- Puzzle Game -->
-            <div class="game-card">
-                <img src="../photos/puzzle1.jpg" alt="Puzzle Game">
-                <h2>Puzzle Game</h2>
-                <p>Rregullo figurat për të zbuluar gjestet! Zbavitje dhe mësim në të njëjtën kohë.</p>
-                <button onclick="window.location.href='puzzle.html'">Luaj Puzzle</button>
-
-            </div>
-            <!-- Memory Game -->
-            <div class="game-card">
-                <img src="../photos/memory.jpg" alt="Memory Game">
-                <h2>Memory Game</h2>
-                <p>Gjej çiftet e gjesteve të njëjta dhe përmirëso kujtesën tënde për gjuhën e shenjave!</p>
-                
-                <button onclick="window.location.href='memory.html'">Luaj Memory</button>
-                
-            </div>
-            
+      <div class="game-container">
+        <!-- Puzzle Game -->
+        <div class="game-card">
+          <img src="../photos/puzzle1.jpg" alt="Puzzle Game" />
+          <h2>Puzzle Game</h2>
+          <p>
+            Rregullo figurat për të zbuluar gjestet! Zbavitje dhe mësim në të
+            njëjtën kohë.
+          </p>
+          <button onclick="window.location.href='puzzle.html'">
+            Luaj Puzzle
+          </button>
         </div>
+        <!-- Memory Game -->
+        <div class="game-card">
+          <img src="../photos/memory.jpg" alt="Memory Game" />
+          <h2>Memory Game</h2>
+          <p>
+            Gjej çiftet e gjesteve të njëjta dhe përmirëso kujtesën tënde për
+            gjuhën e shenjave!
+          </p>
 
- </body>
+          <button onclick="window.location.href='memory.html'">
+            Luaj Memory
+          </button>
+        </div>
+      </div>
+    </div>
+  </body>
 </html>
