+++ src/games/memoryGame/memory.css
@@ -1,11 +1,3 @@
-:root {
-    --deep-blue: #001f3f;
-    --medium-blue: #3a6d8c;
-    --light-blue: #73a7c7;
-    --cream: #ead8b1;
-    --dark-cream: #c7b388;
-  }
-
 
 body {
     font-family: Arial, sans-serif;
