+++ src/games/puzzleGame/puzzle.html
@@ -41,7 +41,7 @@
       <div class="puzzle-pieces" id="puzzle-pieces"></div>
       <div class="puzzle-reference">
         <h3>Reference</h3>
-        <img id="puzzle-image" src="../photos/pz1.jpg" alt="Reference Image">
+        <img id="puzzle-image" src="../../photos/pz1.jpg" alt="Reference Image">
       </div>
     </div>
     <div class="congrats-message hidden">Congratulations! You completed the puzzle!</div>
