+++ webpack.config.js
@@ -105,11 +105,11 @@ const pages = [
   { template: "games/games.html", filename: "games.html", chunks: ["games"] },
 
   {
-    template: "games/puzzle.html",
+    template: "games/puzzleGame/puzzle.html",
     filename: "puzzle.html",
     chunks: ["puzzle"],
   },
-  { template: "games/memory.html", filename: "memory.html", chunks: ["memory"] },
+  { template: "games/memoryGame/memory.html", filename: "memory.html", chunks: ["memory"] },
 
 ];
 
@@ -141,9 +141,8 @@ const entryPoints = {
   userHistory: "./src/profile/userHistory/userHistory.js",
   quiz: "./src/quiz/quiz.js",
   games:"./src/games/games.js",
-  puzzle:"./src/games/puzzle.js",
-  memory:"./src/games/memory.js"
-
+  puzzle:"./src/games/puzzleGame/puzzle.js",
+  memory:"./src/games/memoryGame/memory.js"
 };
 
 module.exports = {
